% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_nino.R
\docType{data}
\name{el_nino}
\alias{el_nino}
\title{El Nino}
\format{
An object of class \code{"character"}.
}
\usage{
el_nino
}
\description{
This dataset consists of 495 annual observations between 1525 to 2020, with 0 representing the absence of an El Nino event and 1 indicating its presence.
El Nino is one of the most influential natural climate patterns on earth. It impacts ocean temperatures, the strength of ocean currents, and the local weather in South America.
}
\references{
{W.H. Quinn, V.T. Neal, and S.E. Antunez De Mayolo. El Nino occurrences over the past four and a half centuries. Journal of Geophysical Research: Oceans, 92(C13):14449–14461, 1987.}
(\href{https://agupubs.onlinelibrary.wiley.com/doi/10.1029/JC092iC13p14449}{Quinn})

{}
(\href{https://www.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php}{el_nino})
}
\keyword{datasets}
